refactor: StorybookControllerDocs를 controller/docs 패키지로 이동
